fix: 修复无法多连接、无法创建回复、无法连接到 minio 的问题

+++ service/impl/PostServiceImpl.cs
@@ -83,7 +83,7 @@ public class PostServiceImpl : PostService
             UpdateTime = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss")
         };
         _dbContext.Replies.Add(reply);
-        Post tPost = await _dbContext.Posts.FindAsync(postId);
+        Post tPost = await _dbContext.Posts.FindAsync((long)postId);
         tPost.UpdateTime = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss");
         await _dbContext.SaveChangesAsync();
         return ResponseResult<Reply?>.Success(ResultCode.REPLY_PUBLISH_SUCCESS, reply);
